+++ pyinaturalist/v1/observations.py
@@ -188,7 +188,7 @@ def get_observation_histogram(**params) -> HistogramResponse:
         'week of year' intervals, and :py:class:`~datetime.datetime` objects for all other intervals.
     """
     response = get(f'{API_V1}/observations/histogram', **params)
-    return convert_observation_histogram(response.json())
+    return convert_observation_histogram(response.json()['results'])
 
 
 @document_request_params(*docs._get_observations, docs._pagination)